feat: sortCommentList 유틸 함수로 분리 (FE-22)
