+++ src/components/board-details/ArticleCommentList/index.tsx
@@ -1,6 +1,6 @@
-import { ArticleComment } from "@coworkers-types";
 import { useQuery } from "@tanstack/react-query";
 import Comment from "@components/commons/Comment";
+import { sortCommentList } from "@utils/sortCommentList";
 import { getArticleComments } from "@api/articleCommentApi";
 
 export default function ArticleCommentList({ articleId }: { articleId: number }) {
@@ -8,17 +8,6 @@ export default function ArticleCommentList({ articleId }: { articleId: number })
     queryKey: ["articleComments", articleId],
     queryFn: () => getArticleComments(articleId, 3),
   });
-
-  const sortCommentList = (comments: ArticleComment[] | undefined) => {
-    if (comments) {
-      const sortedComments = [...comments].sort(
-        (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-      );
-
-      return sortedComments;
-    }
-    return undefined;
-  };
 
   return (
     <div className="flex flex-col gap-16">
