+++ src/components/task-list/TaskCommentList/index.tsx
@@ -1,6 +1,6 @@
-import { Comments } from "@coworkers-types";
 import { useQuery } from "@tanstack/react-query";
 import Comment from "@components/commons/Comment";
+import { sortCommentList } from "@utils/sortCommentList";
 import { getComments } from "@api/commentApi";
 
 export default function TaskCommentList({ taskId }: { taskId: number }) {
@@ -8,17 +8,6 @@ export default function TaskCommentList({ taskId }: { taskId: number }) {
     queryKey: ["taskComments", taskId],
     queryFn: () => getComments(taskId),
   });
-
-  const sortCommentList = (comments: Comments | undefined) => {
-    if (comments) {
-      const sortedComments = [...comments].sort(
-        (a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime()
-      );
-
-      return sortedComments;
-    }
-    return undefined;
-  };
 
   return (
     <div className="flex flex-col gap-16">
